Move connectors into their own module/sub directory
